REFACTOR: move classes to subpackages
