+++ src/main/java/segelzwerg/sporttooolbox/IUnits/Distance.java
@@ -1,6 +1,7 @@
 package segelzwerg.sporttooolbox.IUnits;
 
 import lombok.EqualsAndHashCode;
+import segelzwerg.sporttooolbox.iunits.speed.Speed;
 
 /**
  * Represents distance in kilometers and meters
