+++ src/main/java/segelzwerg/sporttooolbox/calculators/SpeedCalculator.java
@@ -3,8 +3,8 @@ package segelzwerg.sporttooolbox.calculators;
 import lombok.EqualsAndHashCode;
 import segelzwerg.sporttooolbox.IUnits.Distance;
 import segelzwerg.sporttooolbox.IUnits.Pace;
-import segelzwerg.sporttooolbox.IUnits.Speed;
 import segelzwerg.sporttooolbox.IUnits.Time;
+import segelzwerg.sporttooolbox.iunits.speed.Speed;
 
 /**
  * Class to calculate speed
