+++ src/main/java/segelzwerg/sporttooolbox/iunits/speed/SpeedFactory.java
@@ -1,9 +1,4 @@
-package segelzwerg.sporttooolbox.IUnits;
-
-import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
-import segelzwerg.sporttooolbox.iunits.speed.Knot;
-import segelzwerg.sporttooolbox.iunits.speed.MeterPerSecond;
-import segelzwerg.sporttooolbox.iunits.speed.MilePerHour;
+package segelzwerg.sporttooolbox.iunits.speed;
 
 import java.util.Objects;
