+++ src/main/java/segelzwerg/sporttooolbox/web/speed/SpeedPresenter.java
@@ -2,10 +2,10 @@ package segelzwerg.sporttooolbox.web.speed;
 
 import lombok.Getter;
 import lombok.ToString;
-import segelzwerg.sporttooolbox.IUnits.Speed;
 import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
 import segelzwerg.sporttooolbox.iunits.speed.Knot;
 import segelzwerg.sporttooolbox.iunits.speed.MilePerHour;
+import segelzwerg.sporttooolbox.iunits.speed.Speed;
 
 /**
  * Represent speed in different units
