+++ src/test/java/segelzwerg/sporttooolbox/iunits/speed/KnotTest.java
@@ -3,7 +3,6 @@ package segelzwerg.sporttooolbox.iunits.speed;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import segelzwerg.sporttooolbox.IUnits.Distance;
-import segelzwerg.sporttooolbox.IUnits.Speed;
 import segelzwerg.sporttooolbox.IUnits.Time;
 
 import static org.hamcrest.MatcherAssert.assertThat;
