+++ src/test/java/segelzwerg/sporttooolbox/iunits/speed/SpeedFactoryTest.java
@@ -1,11 +1,7 @@
-package segelzwerg.sporttooolbox.IUnits;
+package segelzwerg.sporttooolbox.iunits.speed;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import segelzwerg.sporttooolbox.iunits.speed.KilometerPerHour;
-import segelzwerg.sporttooolbox.iunits.speed.Knot;
-import segelzwerg.sporttooolbox.iunits.speed.MeterPerSecond;
-import segelzwerg.sporttooolbox.iunits.speed.MilePerHour;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.core.IsEqual.equalTo;
